organized executor tests
