+++ .github/workflows/test_on_push.yml
@@ -128,11 +128,6 @@ jobs:
         run: |
           if [ "$RUNNER_OS" == "Linux" ]; then
             env PYTHONPATH=./ pytest tests/unit/executor/ -x
-            env PYTHONPATH=./ pytest tests/unit/test_project_structure.py -x
-            env PYTHONPATH=./ pytest tests/unit/test_predictor_params.py -x
-            env PYTHONPATH=./ pytest tests/unit/test_mongodb_handler.py -x
-            env PYTHONPATH=./ pytest tests/unit/test_mongodb_server.py -x
-            env PYTHONPATH=./ pytest tests/unit/test_cache.py -x
             env PYTHONPATH=./ pytest tests/unit/test_llm_utils.py
             env PYTHONPATH=./ pytest tests/unit/ml_handlers/test_minds_endpoint.py
             env PYTHONPATH=./ pytest tests/unit/ml_handlers/test_openai.py
